+++ example.nix
@@ -1,0 +1,7 @@
+{ pkgs ? import <nixpkgs> { } }:
+
+let
+  nix-hs =
+    import (fetchGit "https://github.com/pjones/nix-hs.git") { inherit pkgs; };
+
+in nix-hs { cabal = ./test/hello-world/hello-world.cabal; }